macos: move TerminalTabColor to its own file
