+++ macos/Sources/Features/Terminal/TerminalTabColor.swift
@@ -1,0 +1,110 @@
+import AppKit
+
+enum TerminalTabColor: Int, CaseIterable {
+    case none
+    case blue
+    case purple
+    case pink
+    case red
+    case orange
+    case yellow
+    case green
+    case teal
+    case graphite
+
+    static let paletteRows: [[TerminalTabColor]] = [
+        [.none, .blue, .purple, .pink, .red],
+        [.orange, .yellow, .green, .teal, .graphite],
+    ]
+
+    var localizedName: String {
+        switch self {
+        case .none:
+            return "None"
+        case .blue:
+            return "Blue"
+        case .purple:
+            return "Purple"
+        case .pink:
+            return "Pink"
+        case .red:
+            return "Red"
+        case .orange:
+            return "Orange"
+        case .yellow:
+            return "Yellow"
+        case .green:
+            return "Green"
+        case .teal:
+            return "Teal"
+        case .graphite:
+            return "Graphite"
+        }
+    }
+
+    var displayColor: NSColor? {
+        switch self {
+        case .none:
+            return nil
+        case .blue:
+            return .systemBlue
+        case .purple:
+            return .systemPurple
+        case .pink:
+            return .systemPink
+        case .red:
+            return .systemRed
+        case .orange:
+            return .systemOrange
+        case .yellow:
+            return .systemYellow
+        case .green:
+            return .systemGreen
+        case .teal:
+            if #available(macOS 13.0, *) {
+                return .systemMint
+            } else {
+                return .systemTeal
+            }
+        case .graphite:
+            return .systemGray
+        }
+    }
+
+    func swatchImage(selected: Bool) -> NSImage {
+        let size = NSSize(width: 18, height: 18)
+        return NSImage(size: size, flipped: false) { rect in
+            let circleRect = rect.insetBy(dx: 1, dy: 1)
+            let circlePath = NSBezierPath(ovalIn: circleRect)
+
+            if let fillColor = self.displayColor {
+                fillColor.setFill()
+                circlePath.fill()
+            } else {
+                NSColor.clear.setFill()
+                circlePath.fill()
+                NSColor.quaternaryLabelColor.setStroke()
+                circlePath.lineWidth = 1
+                circlePath.stroke()
+            }
+
+            if self == .none {
+                let slash = NSBezierPath()
+                slash.move(to: NSPoint(x: circleRect.minX + 2, y: circleRect.minY + 2))
+                slash.line(to: NSPoint(x: circleRect.maxX - 2, y: circleRect.maxY - 2))
+                slash.lineWidth = 1.5
+                NSColor.secondaryLabelColor.setStroke()
+                slash.stroke()
+            }
+
+            if selected {
+                let highlight = NSBezierPath(ovalIn: rect.insetBy(dx: 0.5, dy: 0.5))
+                highlight.lineWidth = 2
+                NSColor.controlAccentColor.setStroke()
+                highlight.stroke()
+            }
+
+            return true
+        }
+    }
+}
